update framework, and test passed
