--- src/graphatoms/system/atoms/__init__.py
@@ -1,7 +1,0 @@
-"""The defination of the basic atoms for the system (like ase.Atoms)."""
-
-from ._box import Box
-from ._eng import Energetics
-from ._struct import Matter, Structure
-
-__all__ = ["Box", "Energetics", "Matter", "Structure"]
