--- src/tests/test_helper.py
@@ -1,550 +1,0 @@
-# ruff: noqa: F401, E501, D103
-
-from io import StringIO
-from time import perf_counter
-
-import numpy as np
-import pytest
-from ase import Atoms
-from ase.io import read
-
-from graphatoms.system.helper import analysis, analysis_site
-
-
-def test_analysis(atoms: Atoms) -> None:
-    """Test analysis of `ase.Atoms`."""
-    assert isinstance(atoms, Atoms)
-    print()
-
-    # test for `analysis`
-    start = perf_counter()
-    sys = analysis(atoms)
-    print(f"time: {perf_counter() - start:.4f}")
-    assert sys.is_outer is not None
-    assert sys.hashes is not None
-    assert np.sum(sys.is_outer) == 200
-    assert len(np.unique(sys.hashes)) == 3
-
-    # test for `analysis_site`
-    start = perf_counter()
-    sites = analysis_site(sys)
-    print(f"time: {perf_counter() - start:.4f}")
-    assert len(sites) == 3
-
-    # from ase.visualize import view
-    # lst: list[Atoms] = []
-    # for site in sites:
-    #     atoms0 = atoms.copy()
-    #     atoms0.numbers[site] = 17
-    #     lst.append(atoms0)
-    # view(lst)
-
-
-@pytest.fixture(scope="module")
-def atoms() -> Atoms:
-    f = StringIO("""500
-Lattice="27.220577427065088 -4.1129194557560175e-09 -8.761550392124388e-11 -3.561892733022198e-09 23.573711576112313 -4.72835056099752e-09 -1.2419683130405167e-10 -7.745050184627368e-09 38.99535731063594" Properties=species:S:1:pos:R:3:tags:I:1:energies:R:1:forces:R:3 energy=-2765.4877452441833 free_energy=-2765.4877452441833 stress="0.000728291917254507 4.1763071329811163e-10 1.9063700420095312e-10 4.1763071329811163e-10 0.0007282914854166791 -5.649601547757633e-10 1.9063700420095312e-10 -5.649601547757633e-10 8.386125672131058e-06" pbc="T T T"
-Pd       0.00000054       1.57157991      14.92963934        5      -5.26092180       0.00000639      -0.00000153      -0.01194571
-Pd       2.72205915       1.57158308      14.92964144        5      -5.26092248      -0.00001225      -0.00003139      -0.01196393
-Pd       5.44411730       1.57158064      14.92963910        5      -5.26092193      -0.00001710      -0.00000113      -0.01196340
-Pd       8.16617360       1.57158075      14.92964224        5      -5.26092349       0.00000404      -0.00000369      -0.01196873
-Pd      10.88822997       1.57158169      14.92964043        5      -5.26092254       0.00000175      -0.00000632      -0.01196116
-Pd      13.61028830       1.57157916      14.92964078        5      -5.26092337       0.00000269       0.00001375      -0.01195944
-Pd      16.33234680       1.57158039      14.92963995        5      -5.26092143       0.00000190       0.00000317      -0.01195963
-Pd      19.05440416       1.57158258      14.92964170        5      -5.26092184       0.00000677      -0.00001540      -0.01196158
-Pd      21.77646092       1.57158100      14.92963908        5      -5.26092006      -0.00000179      -0.00000501      -0.01194890
-Pd      24.49851981       1.57158256      14.92964117        5      -5.26092194       0.00000118      -0.00000619      -0.01197075
-Pd       1.36102978       3.92895245      14.92964022        5      -5.26092215      -0.00001568      -0.00000409      -0.01196067
-Pd       4.08308468       3.92895037      14.92964252        5      -5.26092215       0.00003136       0.00001958      -0.01197246
-Pd       6.80514479       3.92895288      14.92964058        5      -5.26092343      -0.00001039      -0.00000034      -0.01196454
-Pd       9.52720300       3.92895255      14.92964175        5      -5.26092102      -0.00000951      -0.00000826      -0.01197248
-Pd      12.24925825       3.92894982      14.92964018        5      -5.26092205       0.00001694       0.00002098      -0.01195882
-Pd      14.97131773       3.92895113      14.92963980        5      -5.26092080      -0.00000056       0.00000622      -0.01195551
-Pd      17.69337727       3.92895312      14.92964363        5      -5.26092295      -0.00002698      -0.00001679      -0.01197922
-Pd      20.41543191       3.92895341      14.92963916        5      -5.26092157       0.00001585      -0.00000912      -0.01195506
-Pd      23.13749041       3.92895024      14.92964421        5      -5.26092363       0.00000456       0.00001589      -0.01197342
-Pd      25.85954968       3.92895161      14.92964055        5      -5.26092224      -0.00000478       0.00001161      -0.01196374
-Pd       0.00000047       6.28632277      14.92964105        5      -5.26092228      -0.00000881       0.00000766      -0.01195350
-Pd       2.72205827       6.28632200      14.92964054        5      -5.26092282      -0.00000105       0.00000507      -0.01196511
-Pd       5.44411662       6.28632268      14.92964062        5      -5.26092201      -0.00000448      -0.00000483      -0.01194788
-Pd       8.16617390       6.28632313      14.92964094        5      -5.26092291      -0.00000677       0.00000338      -0.01196186
-Pd      10.88822910       6.28632147      14.92964131        5      -5.26092130       0.00002294       0.00001068      -0.01196251
-Pd      13.61028909       6.28632365      14.92964056        5      -5.26092225      -0.00000069      -0.00000736      -0.01194784
-Pd      16.33234806       6.28632229      14.92963957        5      -5.26092149      -0.00003006       0.00000635      -0.01195228
-Pd      19.05440069       6.28632150      14.92963690        5      -5.26092143       0.00003905       0.00002521      -0.01195079
-Pd      21.77646406       6.28632244      14.92964199        5      -5.26092241      -0.00002553       0.00000216      -0.01196982
-Pd      24.49851995       6.28632421      14.92963820        5      -5.26092099       0.00001167      -0.00001886      -0.01194969
-Pd       1.36102807       8.64369343      14.92963732        5      -5.26092001       0.00000139       0.00000555      -0.01194387
-Pd       4.08308672       8.64369288      14.92963870        5      -5.26092175       0.00000202       0.00000891      -0.01194805
-Pd       6.80514568       8.64369285      14.92963807        5      -5.26092141      -0.00001049       0.00000547      -0.01193522
-Pd       9.52720259       8.64369550      14.92963833        5      -5.26092105      -0.00000161      -0.00001297      -0.01194778
-Pd      12.24925962       8.64369297      14.92964009        5      -5.26092178       0.00000405       0.00000556      -0.01195206
-Pd      14.97131843       8.64369366      14.92963717        5      -5.26092140      -0.00000445      -0.00000098      -0.01193972
-Pd      17.69337680       8.64369308      14.92964101        5      -5.26092306      -0.00000557       0.00000802      -0.01195931
-Pd      20.41543370       8.64369519      14.92964325        5      -5.26092275      -0.00000332      -0.00000861      -0.01197367
-Pd      23.13749124       8.64369329      14.92964268        5      -5.26092061      -0.00001367       0.00001188      -0.01197939
-Pd      25.85954654       8.64369709      14.92964073        5      -5.26092315       0.00002822      -0.00002806      -0.01196434
-Pd       0.00000046      11.00106581      14.92964107        5      -5.26092263      -0.00000236      -0.00000184      -0.01195795
-Pd       2.72205779      11.00106600      14.92964233        5      -5.26092163      -0.00000026      -0.00001310      -0.01197486
-Pd       5.44411685      11.00106481      14.92964128        5      -5.26092267      -0.00000273      -0.00000613      -0.01196290
-Pd       8.16617419      11.00106457      14.92964165        5      -5.26092271      -0.00001438       0.00000788      -0.01195903
-Pd      10.88822977      11.00106392      14.92963805        5      -5.26092058       0.00002182       0.00001312      -0.01193906
-Pd      13.61028871      11.00106720      14.92964187        5      -5.26092220      -0.00001234      -0.00001808      -0.01197013
-Pd      16.33234772      11.00106495      14.92963856        5      -5.26092161      -0.00000375      -0.00000332      -0.01195334
-Pd      19.05440268      11.00106497      14.92963691        5      -5.26092066       0.00001642       0.00000527      -0.01193702
-Pd      21.77646242      11.00106754      14.92964149        5      -5.26092260      -0.00001280      -0.00000825      -0.01197936
-Pd      24.49852080      11.00106376      14.92964146        5      -5.26092034      -0.00000235       0.00002483      -0.01197489
-Pd       1.36102812      13.35843682      14.92963864        5      -5.26092110       0.00000299       0.00000066      -0.01196274
-Pd       4.08308633      13.35843716      14.92963993        5      -5.26092125       0.00000950       0.00000222      -0.01196219
-Pd       6.80514310      13.35843570      14.92963695        5      -5.26092127       0.00001402       0.00000287      -0.01193910
-Pd       9.52720208      13.35843579      14.92963885        5      -5.26092045      -0.00000587       0.00000608      -0.01194918
-Pd      12.24925936      13.35843852      14.92963695        5      -5.26092085      -0.00000256      -0.00001790      -0.01193530
-Pd      14.97131636      13.35843507      14.92964156        5      -5.26092077       0.00001215       0.00001484      -0.01196127
-Pd      17.69337539      13.35843755      14.92963834        5      -5.26092202      -0.00001105      -0.00001495      -0.01194669
-Pd      20.41543135      13.35843809      14.92963785        5      -5.26092071       0.00000925      -0.00001616      -0.01194740
-Pd      23.13749123      13.35843718      14.92964244        5      -5.26092222      -0.00000200      -0.00000521      -0.01197082
-Pd      25.85954788      13.35843691      14.92964069        5      -5.26092160       0.00000239      -0.00001203      -0.01196205
-Pd      -0.00000034      15.71580822      14.92963932        5      -5.26092164       0.00000775       0.00000268      -0.01195893
-Pd       2.72205743      15.71580682      14.92963975        5      -5.26092163       0.00000090       0.00001164      -0.01195209
-Pd       5.44411616      15.71580800      14.92963882        5      -5.26092071      -0.00000744       0.00000117      -0.01194962
-Pd       8.16617327      15.71580743      14.92963939        5      -5.26092063      -0.00000268      -0.00000891      -0.01195445
-Pd      10.88823064      15.71580777      14.92963897        5      -5.26092182       0.00000311      -0.00000318      -0.01195623
-Pd      13.61028997      15.71580608      14.92964002        5      -5.26092145      -0.00000860       0.00001049      -0.01196003
-Pd      16.33234798      15.71580644      14.92964094        5      -5.26092285      -0.00001074       0.00000035      -0.01196614
-Pd      19.05440422      15.71580712      14.92963825        5      -5.26092217       0.00000821       0.00001134      -0.01195390
-Pd      21.77646290      15.71580680      14.92963873        5      -5.26092181      -0.00000766       0.00001061      -0.01195082
-Pd      24.49851969      15.71580632      14.92963960        5      -5.26092175      -0.00000602       0.00000948      -0.01195511
-Pd       1.36102958      18.07318002      14.92963887        5      -5.26092140      -0.00000922      -0.00000644      -0.01194999
-Pd       4.08308643      18.07317911      14.92963976        5      -5.26092274       0.00000675      -0.00000627      -0.01195141
-Pd       6.80514489      18.07317909      14.92963982        5      -5.26092217      -0.00001113       0.00000468      -0.01195317
-Pd       9.52720072      18.07317914      14.92964044        5      -5.26092128       0.00001305       0.00000511      -0.01196159
-Pd      12.24925832      18.07317916      14.92964018        5      -5.26092242       0.00001179      -0.00000143      -0.01195765
-Pd      14.97131844      18.07317699      14.92964035        5      -5.26092158      -0.00000694       0.00001891      -0.01195271
-Pd      17.69337504      18.07317760      14.92963980        5      -5.26092112       0.00000651       0.00001796      -0.01195234
-Pd      20.41543311      18.07317903      14.92963980        5      -5.26092139      -0.00000897       0.00000549      -0.01195853
-Pd      23.13749070      18.07317906      14.92963901        5      -5.26092146      -0.00000349      -0.00000894      -0.01195426
-Pd      25.85954739      18.07317766      14.92964111        5      -5.26092245       0.00001518       0.00000815      -0.01196919
-Pd      -0.00000128      20.43055284      14.92964115        5      -5.26092205       0.00001442      -0.00001920      -0.01195568
-Pd       2.72205840      20.43054890      14.92963910        5      -5.26092148      -0.00001326       0.00000480      -0.01194671
-Pd       5.44411406      20.43055154      14.92963779        5      -5.26092073       0.00001890      -0.00001208      -0.01194221
-Pd       8.16617464      20.43055150      14.92964016        5      -5.26092143      -0.00001859       0.00000109      -0.01196220
-Pd      10.88823139      20.43055052      14.92964161        5      -5.26092246       0.00000601      -0.00000729      -0.01196727
-Pd      13.61028882      20.43054995      14.92964137        5      -5.26092154       0.00000298      -0.00000091      -0.01194927
-Pd      16.33234614      20.43055177      14.92963764        5      -5.26092127      -0.00001201      -0.00001431      -0.01194055
-Pd      19.05440156      20.43055260      14.92964100        5      -5.26092236       0.00002170      -0.00003428      -0.01196103
-Pd      21.77646307      20.43054983      14.92963889        5      -5.26092106      -0.00000550       0.00000510      -0.01195076
-Pd      24.49852174      20.43054788      14.92963920        5      -5.26092179      -0.00001970       0.00001477      -0.01195496
-Pd       1.36103041      22.78792002      14.92963849        5      -5.26092069      -0.00001273       0.00001460      -0.01193959
-Pd       4.08308399      22.78791898      14.92963790        5      -5.26092077       0.00001005       0.00002350      -0.01194765
-Pd       6.80514262      22.78792232      14.92963753        5      -5.26092029       0.00002008      -0.00001228      -0.01194823
-Pd       9.52720541      22.78792199      14.92964285        5      -5.26092326      -0.00003284       0.00000288      -0.01197013
-Pd      12.24925898      22.78792262      14.92963977        5      -5.26092177       0.00001186      -0.00001198      -0.01195218
-Pd      14.97131642      22.78792235      14.92963997        5      -5.26092218       0.00000893      -0.00001368      -0.01195398
-Pd      17.69337666      22.78792159      14.92963919        5      -5.26092178      -0.00001636       0.00001006      -0.01196002
-Pd      20.41543272      22.78791948      14.92964033        5      -5.26092214       0.00001393       0.00001800      -0.01196019
-Pd      23.13749222      22.78792231      14.92964065        5      -5.26092179      -0.00001953      -0.00001235      -0.01196344
-Pd      25.85954631      22.78792165      14.92964033        5      -5.26092271       0.00002562       0.00000273      -0.01196639
-Pd      -0.00000040       0.00000026      17.22794424        4      -5.65149231       0.00000905       0.00000727       0.02450913
-Pd       2.72205806       0.00000194      17.22794530        4      -5.65149242      -0.00000100      -0.00001446       0.02450920
-Pd       5.44411687       0.00000118      17.22793922        4      -5.65149250      -0.00001009      -0.00000095       0.02455330
-Pd       8.16617333      -0.00000027      17.22794410        4      -5.65149143       0.00000061       0.00000358       0.02451988
-Pd      10.88823126       0.00000060      17.22794358        4      -5.65149034       0.00000007       0.00000749       0.02452787
-Pd      13.61028961      -0.00000003      17.22794382        4      -5.65149123      -0.00000859       0.00000793       0.02452499
-Pd      16.33234682      -0.00000129      17.22794278        4      -5.65149138       0.00000332       0.00000615       0.02453086
-Pd      19.05440362      -0.00000209      17.22794199        4      -5.65149135       0.00000412       0.00002119       0.02454322
-Pd      21.77646333       0.00000107      17.22794353        4      -5.65149143      -0.00000996       0.00000179       0.02452382
-Pd      24.49852083      -0.00000006      17.22793786        4      -5.65148915      -0.00000729       0.00000788       0.02457649
-Pd       1.36103056       2.35737270      17.22794371        4      -5.65148872      -0.00002164      -0.00002117       0.02451176
-Pd       4.08308578       2.35737317      17.22794051        4      -5.65148917       0.00002254      -0.00002842       0.02454877
-Pd       6.80514472       2.35737205      17.22794002        4      -5.65148954      -0.00001561      -0.00001595       0.02455764
-Pd       9.52720099       2.35737302      17.22794319        4      -5.65148779       0.00001288      -0.00002515       0.02452799
-Pd      12.24925806       2.35737235      17.22794085        4      -5.65149181       0.00001563      -0.00000524       0.02455280
-Pd      14.97131932       2.35737085      17.22794174        4      -5.65149152      -0.00001141       0.00000133       0.02453809
-Pd      17.69337649       2.35737187      17.22794156        4      -5.65148880      -0.00000526      -0.00001009       0.02454446
-Pd      20.41543317       2.35737143      17.22794504        4      -5.65149294       0.00000174      -0.00000033       0.02451926
-Pd      23.13749230       2.35737119      17.22794250        4      -5.65148920      -0.00001529      -0.00000850       0.02454065
-Pd      25.85954764       2.35737417      17.22794244        4      -5.65148851       0.00001373      -0.00003278       0.02452219
-Pd      -0.00000039       4.71474107      17.22794408        4      -5.65149018       0.00000904       0.00002740       0.02451381
-Pd       2.72205800       4.71473965      17.22794204        4      -5.65148926      -0.00000416       0.00003053       0.02453778
-Pd       5.44411573       4.71474019      17.22794625        4      -5.65149111      -0.00000242       0.00001992       0.02450179
-Pd       8.16617462       4.71473963      17.22794152        4      -5.65148798      -0.00001499       0.00002389       0.02453545
-Pd      10.88822895       4.71474250      17.22794509        4      -5.65148939       0.00002977      -0.00000727       0.02451866
-Pd      13.61028939       4.71474166      17.22794156        4      -5.65149298      -0.00000743       0.00000161       0.02453058
-Pd      16.33234726       4.71474407      17.22794345        4      -5.65148881      -0.00001729      -0.00001714       0.02452090
-Pd      19.05440290       4.71474402      17.22794073        4      -5.65148823       0.00002019      -0.00000448       0.02455045
-Pd      21.77646104       4.71474326      17.22794192        4      -5.65148916       0.00000116      -0.00000226       0.02453621
-Pd      24.49852062       4.71474142      17.22794434        4      -5.65148936      -0.00000965       0.00001790       0.02450870
-Pd       1.36103083       7.07211323      17.22794278        4      -5.65149271      -0.00002236      -0.00000046       0.02452862
-Pd       4.08308384       7.07211183      17.22794202        4      -5.65149231       0.00003334       0.00000599       0.02453686
-Pd       6.80514653       7.07211143      17.22794525        4      -5.65149277      -0.00002494       0.00002304       0.02450756
-Pd       9.52720325       7.07211228      17.22794156        4      -5.65148860      -0.00001745       0.00001416       0.02453273
-Pd      12.24925818       7.07211368      17.22794500        4      -5.65148995       0.00002048      -0.00000631       0.02450874
-Pd      14.97131695       7.07211286      17.22794769        4      -5.65149237      -0.00000686       0.00000472       0.02449398
-Pd      17.69337371       7.07211348      17.22794308        4      -5.65148810       0.00001266       0.00000467       0.02452518
-Pd      20.41543276       7.07211548      17.22794091        4      -5.65149299       0.00000407      -0.00001456       0.02454158
-Pd      23.13749219       7.07211444      17.22794047        4      -5.65149033      -0.00001343      -0.00000327       0.02455507
-Pd      25.85954796       7.07211321      17.22794411        4      -5.65149288       0.00001240      -0.00000009       0.02451895
-Pd      -0.00000063       9.42948389      17.22794373        4      -5.65149337       0.00000960      -0.00000095       0.02452458
-Pd       2.72205715       9.42948556      17.22794326        4      -5.65148950      -0.00000429      -0.00000956       0.02451406
-Pd       5.44411294       9.42948544      17.22794392        4      -5.65149294       0.00002443      -0.00002027       0.02451744
-Pd       8.16617383       9.42948536      17.22794416        4      -5.65149209      -0.00001671      -0.00001154       0.02450999
-Pd      10.88823031       9.42948252      17.22794558        4      -5.65149183       0.00000785       0.00002450       0.02450683
-Pd      13.61028916       9.42948533      17.22794126        4      -5.65149185      -0.00000298      -0.00000895       0.02453034
-Pd      16.33234574       9.42948303      17.22794315        4      -5.65149051       0.00000489       0.00001400       0.02452014
-Pd      19.05440524       9.42948258      17.22794450        4      -5.65149058      -0.00001238       0.00001844       0.02451566
-Pd      21.77646247       9.42948831      17.22794182        4      -5.65148863      -0.00000309      -0.00003825       0.02454492
-Pd      24.49851962       9.42948520      17.22793904        4      -5.65148667       0.00000732      -0.00001038       0.02456487
-Pd       1.36102826      11.78685699      17.22794165        4      -5.65148894      -0.00001125      -0.00001175       0.02454172
-Pd       4.08308455      11.78685688      17.22793808        4      -5.65148864       0.00002159      -0.00000086       0.02457048
-Pd       6.80514291      11.78685724      17.22794477        4      -5.65148963       0.00000453      -0.00001036       0.02451155
-Pd       9.52720153      11.78685588      17.22794181        4      -5.65149256       0.00001149      -0.00000077       0.02453912
-Pd      12.24926093      11.78685638      17.22794467        4      -5.65149065      -0.00001910      -0.00000254       0.02450684
-Pd      14.97131799      11.78685603      17.22793801        4      -5.65149204       0.00001608       0.00000692       0.02456695
-Pd      17.69337704      11.78685444      17.22794343        4      -5.65149355      -0.00001548       0.00001544       0.02452271
-Pd      20.41543591      11.78685428      17.22794421        4      -5.65149075      -0.00003438       0.00002115       0.02451227
-Pd      23.13749142      11.78685306      17.22793840        4      -5.65148769       0.00001009       0.00003306       0.02456624
-Pd      25.85954824      11.78685558      17.22794624        4      -5.65148979       0.00000834      -0.00000019       0.02451383
-Pd      -0.00000028      14.14422579      17.22793983        4      -5.65148998       0.00000943       0.00001077       0.02455793
-Pd       2.72205892      14.14422409      17.22794165        4      -5.65149267      -0.00000311       0.00003092       0.02454812
-Pd       5.44411838      14.14422841      17.22794469        4      -5.65149113      -0.00002975      -0.00001217       0.02451259
-Pd       8.16617274      14.14422675      17.22794575        4      -5.65149148       0.00001807      -0.00000103       0.02450714
-Pd      10.88823012      14.14422581      17.22794309        4      -5.65149208       0.00000674       0.00001575       0.02453006
-Pd      13.61028886      14.14422946      17.22794530        4      -5.65149057      -0.00001166      -0.00001893       0.02450952
-Pd      16.33234552      14.14423065      17.22794438        4      -5.65149339       0.00001032      -0.00004031       0.02453061
-Pd      19.05440273      14.14422752      17.22794089        4      -5.65149278       0.00001535      -0.00000708       0.02454396
-Pd      21.77646256      14.14422624      17.22794139        4      -5.65149230      -0.00000700       0.00001178       0.02454103
-Pd      24.49852152      14.14422649      17.22794423        4      -5.65148963      -0.00001940       0.00000621       0.02452735
-Pd       1.36102767      16.50159716      17.22794387        4      -5.65149216       0.00000987      -0.00000327       0.02452242
-Pd       4.08308529      16.50159767      17.22794366        4      -5.65149201      -0.00000302      -0.00000374       0.02451161
-Pd       6.80514435      16.50159604      17.22794352        4      -5.65149008       0.00000701       0.00002644       0.02451279
-Pd       9.52720274      16.50159935      17.22793986        4      -5.65149387      -0.00000078      -0.00001460       0.02455748
-Pd      12.24925951      16.50160026      17.22794099        4      -5.65149052      -0.00000754      -0.00001620       0.02454103
-Pd      14.97131726      16.50159927      17.22794163        4      -5.65148867       0.00000322       0.00000942       0.02453855
-Pd      17.69337502      16.50159885      17.22794188        4      -5.65148862       0.00000780      -0.00000158       0.02453958
-Pd      20.41543312      16.50159986      17.22794234        4      -5.65149195      -0.00000224      -0.00002480       0.02452792
-Pd      23.13749040      16.50159896      17.22794379        4      -5.65149163       0.00000139      -0.00000913       0.02453042
-Pd      25.85954782      16.50159862      17.22794092        4      -5.65148970       0.00000132      -0.00000559       0.02454849
-Pd      -0.00000018      18.85896810      17.22794159        4      -5.65149010      -0.00000369       0.00001095       0.02453181
-Pd       2.72205744      18.85896990      17.22794297        4      -5.65149074       0.00000436      -0.00000875       0.02451927
-Pd       5.44411639      18.85896926      17.22794334        4      -5.65149137       0.00000102       0.00000093       0.02451025
-Pd       8.16617335      18.85896937      17.22794188        4      -5.65148969      -0.00000656      -0.00000644       0.02454283
-Pd      10.88823182      18.85897139      17.22794523        4      -5.65148918      -0.00000695      -0.00001913       0.02451004
-Pd      13.61028881      18.85896798      17.22794408        4      -5.65149096       0.00000860       0.00001558       0.02451372
-Pd      16.33234881      18.85896900      17.22794465        4      -5.65149260      -0.00001788       0.00000783       0.02451135
-Pd      19.05440453      18.85896911      17.22794588        4      -5.65148978      -0.00000449       0.00000527       0.02450341
-Pd      21.77645975      18.85896948      17.22794172        4      -5.65149089       0.00002422       0.00000923       0.02453679
-Pd      24.49851963      18.85896920      17.22794188        4      -5.65149290      -0.00001106       0.00000611       0.02455100
-Pd       1.36102800      21.21634049      17.22794535        4      -5.65149167       0.00000218       0.00000564       0.02449922
-Pd       4.08308761      21.21634093      17.22794353        4      -5.65149134      -0.00000332       0.00000252       0.02451285
-Pd       6.80514657      21.21634042      17.22794403        4      -5.65149212      -0.00001329      -0.00000052       0.02450977
-Pd       9.52720328      21.21633988      17.22794021        4      -5.65148783      -0.00001154       0.00000569       0.02455132
-Pd      12.24925822      21.21634108      17.22794492        4      -5.65148988       0.00002324       0.00000332       0.02451434
-Pd      14.97131684      21.21634122      17.22794610        4      -5.65149209       0.00000821      -0.00001339       0.02449939
-Pd      17.69337559      21.21633930      17.22794044        4      -5.65149080      -0.00001471       0.00000683       0.02455085
-Pd      20.41543264      21.21634188      17.22794220        4      -5.65148796       0.00001036      -0.00001110       0.02452543
-Pd      23.13749072      21.21634093      17.22794430        4      -5.65149169      -0.00000237      -0.00001293       0.02451926
-Pd      25.85954611      21.21634158      17.22794248        4      -5.65149032       0.00001434      -0.00001466       0.02454042
-Pd       1.36102727       0.78579323      19.49767888        3      -5.83005228       0.00001230      -0.00002827       0.00000495
-Pd       4.08308583       0.78579151      19.49767826        3      -5.83005540       0.00001757      -0.00001354       0.00000273
-Pd       6.80514673       0.78579016      19.49768101        3      -5.83005487      -0.00003712       0.00000126      -0.00001403
-Pd       9.52720354       0.78579101      19.49767874        3      -5.83005061      -0.00001316      -0.00001817      -0.00000120
-Pd      12.24925942       0.78579016      19.49767886        3      -5.83005231       0.00001348       0.00000651      -0.00001556
-Pd      14.97131636       0.78578998      19.49767969        3      -5.83005135       0.00000505       0.00001429      -0.00001050
-Pd      17.69337294       0.78579181      19.49767934        3      -5.83005492       0.00002428      -0.00002252      -0.00000592
-Pd      20.41543224       0.78578940      19.49768011        3      -5.83005505      -0.00000429       0.00000947      -0.00000722
-Pd      23.13749009       0.78578914      19.49768059        3      -5.83005361       0.00001167       0.00000936      -0.00002131
-Pd      25.85954947       0.78579214      19.49767714        3      -5.83005328      -0.00001870      -0.00000898       0.00000338
-Pd      -0.00000075       3.14316316      19.49767426        3      -5.83005124       0.00000600      -0.00001099       0.00003365
-Pd       2.72205762       3.14316383      19.49767786        3      -5.83005210      -0.00000203      -0.00001918       0.00000292
-Pd       5.44411366       3.14316221      19.49767964        3      -5.83005367       0.00002073      -0.00002932      -0.00000652
-Pd       8.16617246       3.14316062      19.49767659        3      -5.83005327      -0.00001082       0.00001711       0.00001350
-Pd      10.88822795       3.14316002      19.49768101        3      -5.83005176       0.00003155       0.00001690      -0.00001990
-Pd      13.61028984       3.14316126      19.49767862        3      -5.83005562      -0.00001673      -0.00000595      -0.00000815
-Pd      16.33234680       3.14316135      19.49767857        3      -5.83004951      -0.00000514      -0.00000240      -0.00000569
-Pd      19.05440341       3.14316140      19.49767994        3      -5.83005194       0.00001164      -0.00000238      -0.00001210
-Pd      21.77646391       3.14316125      19.49767718        3      -5.83005289      -0.00001700      -0.00000783       0.00000742
-Pd      24.49852018       3.14316158      19.49767910        3      -5.83004937      -0.00000659      -0.00001602      -0.00000753
-Pd       1.36102906       5.50053233      19.49767861        3      -5.83005315      -0.00000607       0.00001117      -0.00001223
-Pd       4.08308647       5.50053189      19.49768103        3      -5.83004942      -0.00000811       0.00001415      -0.00002977
-Pd       6.80514259       5.50053312      19.49767491        3      -5.83005142       0.00002958      -0.00000813       0.00002581
-Pd       9.52720380       5.50053293      19.49767642        3      -5.83005333      -0.00002809      -0.00000056       0.00002680
-Pd      12.24925991       5.50053238      19.49767722        3      -5.83005352       0.00001088      -0.00000819       0.00000587
-Pd      14.97131846       5.50052994      19.49767982        3      -5.83005133      -0.00001882       0.00001644       0.00000054
-Pd      17.69337392       5.50053147      19.49767668        3      -5.83005375       0.00002571       0.00000741       0.00001565
-Pd      20.41543420       5.50053079      19.49767788        3      -5.83005526      -0.00000904       0.00001401       0.00000363
-Pd      23.13749258       5.50053199      19.49767619        3      -5.83005091      -0.00002783       0.00001629       0.00000909
-Pd      25.85954619       5.50053314      19.49767907        3      -5.83004991       0.00002358       0.00000617      -0.00000834
-Pd      -0.00000024       7.85790326      19.49768016        3      -5.83005422       0.00000596       0.00000828      -0.00000834
-Pd       2.72205813       7.85790450      19.49767854        3      -5.83005117      -0.00000787      -0.00000660       0.00000235
-Pd       5.44411394       7.85790378      19.49768064        3      -5.83005095       0.00001283       0.00000480      -0.00002005
-Pd       8.16617173       7.85790450      19.49767825        3      -5.83005148       0.00000767      -0.00001243       0.00000642
-Pd      10.88823119       7.85790373      19.49767886        3      -5.83004927      -0.00001409       0.00000100       0.00000105
-Pd      13.61028787       7.85790392      19.49767895        3      -5.83005280       0.00001749      -0.00000668      -0.00000037
-Pd      16.33234708       7.85790436      19.49767918        3      -5.83005055      -0.00001004      -0.00000328      -0.00000083
-Pd      19.05440359       7.85790354      19.49767904        3      -5.83005221       0.00001214      -0.00001752      -0.00000272
-Pd      21.77646109       7.85790292      19.49767960        3      -5.83005476       0.00000300      -0.00000329      -0.00001195
-Pd      24.49851935       7.85790443      19.49767939        3      -5.83005269      -0.00000916       0.00000354      -0.00001185
-Pd       1.36103050      10.21527458      19.49767798        3      -5.83005264      -0.00002807       0.00000184       0.00000875
-Pd       4.08308624      10.21527239      19.49767780        3      -5.83005515       0.00001187       0.00003440       0.00001152
-Pd       6.80514676      10.21527509      19.49767839        3      -5.83004981      -0.00002818      -0.00001227       0.00000652
-Pd       9.52720127      10.21527491      19.49767874        3      -5.83005270       0.00002676       0.00001257       0.00000859
-Pd      12.24926106      10.21527426      19.49767928        3      -5.83005210      -0.00001590       0.00001110       0.00000072
-Pd      14.97131872      10.21527499      19.49767851        3      -5.83005209      -0.00000517       0.00000422      -0.00000680
-Pd      17.69337680      10.21527522      19.49767807        3      -5.83005001      -0.00002087       0.00000411       0.00000520
-Pd      20.41543207      10.21527445      19.49767939        3      -5.83004997       0.00001809       0.00000214       0.00000297
-Pd      23.13748966      10.21527368      19.49767586        3      -5.83005497       0.00000190       0.00002181       0.00001524
-Pd      25.85954676      10.21527500      19.49767828        3      -5.83005414       0.00001771      -0.00000079       0.00000176
-Pd       0.00000022      12.57264652      19.49768156        3      -5.83005384      -0.00000781      -0.00000458      -0.00001719
-Pd       2.72205473      12.57264570      19.49767790        3      -5.83005577       0.00004162       0.00000579       0.00000380
-Pd       5.44411713      12.57264521      19.49767914        3      -5.83005484      -0.00003569       0.00000252      -0.00000331
-Pd       8.16617214      12.57264475      19.49768039        3      -5.83005057       0.00001476       0.00001721      -0.00001320
-Pd      10.88823196      12.57264798      19.49767984        3      -5.83005435      -0.00001474      -0.00001895      -0.00000616
-Pd      13.61028670      12.57264431      19.49767742        3      -5.83005395       0.00002278       0.00001621       0.00001639
-Pd      16.33234386      12.57264575      19.49768284        3      -5.83005444       0.00001261       0.00001116      -0.00003114
-Pd      19.05440336      12.57264538      19.49767854        3      -5.83005535       0.00000097       0.00001932       0.00000406
-Pd      21.77646101      12.57264821      19.49767691        3      -5.83005383       0.00001021      -0.00001820       0.00001958
-Pd      24.49852024      12.57264779      19.49767811        3      -5.83005270      -0.00001611      -0.00001548       0.00000756
-Pd       1.36102887      14.93001631      19.49768102        3      -5.83005437       0.00000049       0.00000136      -0.00002316
-Pd       4.08308884      14.93001860      19.49768166        3      -5.83005154      -0.00002765      -0.00002673      -0.00002511
-Pd       6.80514454      14.93001937      19.49767698        3      -5.83004927       0.00000524      -0.00002193       0.00002051
-Pd       9.52720233      14.93001659      19.49768304        3      -5.83005436       0.00001143       0.00000709      -0.00003795
-Pd      12.24926239      14.93001922      19.49767986        3      -5.83005033      -0.00003307      -0.00001449      -0.00001122
-Pd      14.97131614      14.93001999      19.49767703        3      -5.83005095       0.00002091      -0.00002830       0.00001385
-Pd      17.69337578      14.93001751      19.49767977        3      -5.83005359      -0.00000403      -0.00001292      -0.00001194
-Pd      20.41543313      14.93001722      19.49768017        3      -5.83005395       0.00000034      -0.00000648      -0.00001049
-Pd      23.13749176      14.93001643      19.49768114        3      -5.83005086      -0.00001396       0.00001106      -0.00001840
-Pd      25.85954761      14.93001910      19.49767833        3      -5.83005255       0.00001995      -0.00000961       0.00000035
-Pd       0.00000078      17.28738753      19.49767957        3      -5.83005292      -0.00000516       0.00000158      -0.00000434
-Pd       2.72205723      17.28738804      19.49767956        3      -5.83005041      -0.00000358      -0.00000084      -0.00000555
-Pd       5.44411399      17.28738681      19.49767434        3      -5.83005214       0.00002201       0.00001828       0.00003773
-Pd       8.16617429      17.28738787      19.49768047        3      -5.83005389      -0.00001208       0.00001186      -0.00001903
-Pd      10.88823013      17.28738832      19.49768024        3      -5.83005197       0.00000438      -0.00001367      -0.00001028
-Pd      13.61028774      17.28738792      19.49768042        3      -5.83005238       0.00000375       0.00002123      -0.00001154
-Pd      16.33234747      17.28738825      19.49768129        3      -5.83005308      -0.00000490       0.00000653      -0.00002008
-Pd      19.05440439      17.28738928      19.49767886        3      -5.83005040      -0.00000830      -0.00000356      -0.00000728
-Pd      21.77646003      17.28738820      19.49767952        3      -5.83005066       0.00001899      -0.00000393      -0.00000672
-Pd      24.49852033      17.28738832      19.49768131        3      -5.83005398      -0.00000952       0.00000370      -0.00002420
-Pd       1.36102985      19.64475944      19.49767541        3      -5.83005078      -0.00000617      -0.00000410       0.00002915
-Pd       4.08308878      19.64475693      19.49767742        3      -5.83005149      -0.00001437       0.00001195       0.00001607
-Pd       6.80514441      19.64476039      19.49768041        3      -5.83005306       0.00001257      -0.00001535      -0.00000711
-Pd       9.52720285      19.64475840      19.49767762        3      -5.83005335      -0.00001999       0.00001874       0.00000774
-Pd      12.24925790      19.64475966      19.49767740        3      -5.83005202       0.00001997       0.00000274       0.00002103
-Pd      14.97131697      19.64475968      19.49767637        3      -5.83005348       0.00000585      -0.00000724       0.00002480
-Pd      17.69337519      19.64476182      19.49767962        3      -5.83005436      -0.00000700      -0.00000680       0.00000166
-Pd      20.41543311      19.64476033      19.49767752        3      -5.83005250       0.00000319      -0.00000208       0.00001586
-Pd      23.13748937      19.64475860      19.49768220        3      -5.83005312       0.00001492      -0.00000065      -0.00003425
-Pd      25.85954999      19.64475831      19.49768000        3      -5.83005393      -0.00001230       0.00000719      -0.00001415
-Pd      -0.00000234      22.00213104      19.49767840        3      -5.83005302       0.00001839       0.00000139       0.00000470
-Pd       2.72205739      22.00212887      19.49767669        3      -5.83005404      -0.00000361       0.00002512       0.00002103
-Pd       5.44411575      22.00212946      19.49767718        3      -5.83005514      -0.00000356       0.00002784       0.00001331
-Pd       8.16617311      22.00213171      19.49767766        3      -5.83005495       0.00001285      -0.00001020       0.00002100
-Pd      10.88823044      22.00213147      19.49767746        3      -5.83005387      -0.00000168      -0.00000114       0.00001093
-Pd      13.61029151      22.00213169      19.49767911        3      -5.83005021      -0.00003406      -0.00000895       0.00000183
-Pd      16.33234481      22.00213235      19.49768148        3      -5.83005192       0.00003322      -0.00000755      -0.00002227
-Pd      19.05440635      22.00212931      19.49767863        3      -5.83005390      -0.00003001       0.00002204       0.00000012
-Pd      21.77646308      22.00212978      19.49767749        3      -5.83005061      -0.00000605       0.00000320       0.00000473
-Pd      24.49851857      22.00213000      19.49768041        3      -5.83005290       0.00000917       0.00000611      -0.00002083
-Pd       0.00000260       1.57157866      21.76741692        2      -5.65148880      -0.00003273       0.00002282      -0.02455000
-Pd       2.72205559       1.57158069      21.76741294        2      -5.65148977       0.00002624      -0.00000277      -0.02451839
-Pd       5.44411414       1.57158092      21.76741804        2      -5.65149307       0.00000590      -0.00001029      -0.02455083
-Pd       8.16617422       1.57157988      21.76741343        2      -5.65148889      -0.00000738       0.00001035      -0.02452524
-Pd      10.88823229       1.57157983      21.76740981        2      -5.65148897      -0.00000207       0.00000894      -0.02449584
-Pd      13.61028948       1.57157982      21.76741531        2      -5.65149014      -0.00000409       0.00001009      -0.02452627
-Pd      16.33234641       1.57158271      21.76741342        2      -5.65148838      -0.00000216      -0.00002344      -0.02452190
-Pd      19.05440224       1.57158022      21.76741631        2      -5.65148905       0.00001010      -0.00001116      -0.02453516
-Pd      21.77646071       1.57158221      21.76741682        2      -5.65148883      -0.00000022       0.00000370      -0.02454853
-Pd      24.49851935       1.57158057      21.76741353        2      -5.65148964       0.00001268       0.00000228      -0.02452311
-Pd       1.36102960       3.92895206      21.76741299        2      -5.65149215      -0.00001079      -0.00000225      -0.02451368
-Pd       4.08308694       3.92895083      21.76741254        2      -5.65149095      -0.00000840       0.00000116      -0.02450048
-Pd       6.80514440       3.92895153      21.76741335        2      -5.65149229       0.00000630       0.00000368      -0.02452007
-Pd       9.52720139       3.92894981      21.76741475        2      -5.65149265       0.00000347       0.00002427      -0.02452914
-Pd      12.24926069       3.92895335      21.76741231        2      -5.65149160      -0.00000501      -0.00000774      -0.02451158
-Pd      14.97131843       3.92895071      21.76741660        2      -5.65148967      -0.00000926       0.00001303      -0.02453935
-Pd      17.69337582       3.92894986      21.76741539        2      -5.65148903      -0.00000083       0.00002216      -0.02453714
-Pd      20.41543104       3.92895146      21.76741351        2      -5.65149360       0.00002155       0.00001009      -0.02452044
-Pd      23.13749144       3.92895072      21.76741160        2      -5.65149305      -0.00001580       0.00001003      -0.02450009
-Pd      25.85954900       3.92895034      21.76740963        2      -5.65149317       0.00001163      -0.00000176      -0.02448355
-Pd       0.00000058       6.28632173      21.76741261        2      -5.65149307      -0.00000386       0.00000656      -0.02449796
-Pd       2.72205856       6.28632288      21.76741105        2      -5.65149089      -0.00001332       0.00000012      -0.02449369
-Pd       5.44411419       6.28632201      21.76741159        2      -5.65148915       0.00001788      -0.00000463      -0.02450360
-Pd       8.16617187       6.28632255      21.76741355        2      -5.65149426       0.00001490      -0.00000610      -0.02452325
-Pd      10.88823342       6.28632411      21.76741340        2      -5.65149235      -0.00003137      -0.00001570      -0.02451861
-Pd      13.61028877       6.28632446      21.76741560        2      -5.65148888       0.00001247      -0.00001255      -0.02454057
-Pd      16.33234666       6.28632152      21.76741104        2      -5.65149278      -0.00000155       0.00000738      -0.02450569
-Pd      19.05440372       6.28632440      21.76741527        2      -5.65149032       0.00000125      -0.00001954      -0.02454607
-Pd      21.77646229       6.28632420      21.76741631        2      -5.65149202       0.00000593      -0.00001772      -0.02454533
-Pd      24.49851956       6.28632169      21.76741198        2      -5.65149182      -0.00000731       0.00001036      -0.02449778
-Pd       1.36102914       8.64369444      21.76741564        2      -5.65149170      -0.00000212      -0.00000162      -0.02453345
-Pd       4.08308916       8.64369219      21.76741664        2      -5.65148845      -0.00002467       0.00002672      -0.02453801
-Pd       6.80514351       8.64369411      21.76741688        2      -5.65148866       0.00001194       0.00000500      -0.02454144
-Pd       9.52720111       8.64369393      21.76741264        2      -5.65149288       0.00000306       0.00000161      -0.02450737
-Pd      12.24925976       8.64369282      21.76741135        2      -5.65148973      -0.00000155       0.00001179      -0.02450149
-Pd      14.97131841       8.64369592      21.76741452        2      -5.65148983      -0.00000686      -0.00001788      -0.02452247
-Pd      17.69337561       8.64369281      21.76741213        2      -5.65149279      -0.00000449       0.00001671      -0.02450286
-Pd      20.41543263       8.64369430      21.76741660        2      -5.65148773       0.00000364       0.00000163      -0.02454695
-Pd      23.13748963       8.64369618      21.76741516        2      -5.65149125       0.00000757      -0.00001284      -0.02452893
-Pd      25.85954681       8.64369431      21.76741301        2      -5.65149031       0.00001642      -0.00000506      -0.02451310
-Pd       0.00000030      11.00106583      21.76741705        2      -5.65148862       0.00000197      -0.00001941      -0.02454779
-Pd       2.72205739      11.00106643      21.76741397        2      -5.65149307       0.00001703      -0.00000690      -0.02453268
-Pd       5.44411645      11.00106670      21.76741401        2      -5.65149347      -0.00001596      -0.00000809      -0.02452928
-Pd       8.16617372      11.00106802      21.76741388        2      -5.65148942      -0.00000058      -0.00003061      -0.02451777
-Pd      10.88823206      11.00106543      21.76741838        2      -5.65148897      -0.00001408      -0.00000896      -0.02456459
-Pd      13.61028774      11.00106465      21.76741663        2      -5.65148937       0.00000633       0.00000642      -0.02455168
-Pd      16.33234378      11.00106505      21.76741589        2      -5.65148939       0.00003186      -0.00000189      -0.02453752
-Pd      19.05440583      11.00106442      21.76741255        2      -5.65149124      -0.00002310       0.00000774      -0.02451541
-Pd      21.77646270      11.00106557      21.76741585        2      -5.65149217       0.00000351       0.00000550      -0.02454792
-Pd      24.49852176      11.00106502      21.76741283        2      -5.65149084      -0.00001678       0.00000388      -0.02452636
-Pd       1.36102705      13.35843588      21.76741631        2      -5.65148824       0.00002382      -0.00000407      -0.02454052
-Pd       4.08308695      13.35843919      21.76741483        2      -5.65148901      -0.00001009      -0.00003047      -0.02454258
-Pd       6.80514492      13.35843630      21.76741674        2      -5.65148616      -0.00001165       0.00000056      -0.02455904
-Pd       9.52720004      13.35843630      21.76741405        2      -5.65148765       0.00001823      -0.00000028      -0.02453290
-Pd      12.24925806      13.35843534      21.76741565        2      -5.65148836       0.00001836       0.00000550      -0.02454509
-Pd      14.97131899      13.35843708      21.76741667        2      -5.65148968      -0.00002351      -0.00000386      -0.02455820
-Pd      17.69337543      13.35843781      21.76741812        2      -5.65148961       0.00000662      -0.00002313      -0.02455681
-Pd      20.41543138      13.35843636      21.76741949        2      -5.65149033       0.00001531       0.00000482      -0.02456930
-Pd      23.13749135      13.35843524      21.76741555        2      -5.65149075      -0.00000227      -0.00000208      -0.02453930
-Pd      25.85955067      13.35843675      21.76741593        2      -5.65148868      -0.00002743      -0.00000201      -0.02454403
-Pd       0.00000034      15.71580665      21.76741825        2      -5.65148833      -0.00000903       0.00001200      -0.02456272
-Pd       2.72205605      15.71580669      21.76741430        2      -5.65149063       0.00001343       0.00000765      -0.02452163
-Pd       5.44411607      15.71580510      21.76741462        2      -5.65148946       0.00000059       0.00003878      -0.02452784
-Pd       8.16617270      15.71580618      21.76741206        2      -5.65149093       0.00000323       0.00002502      -0.02450583
-Pd      10.88823142      15.71580600      21.76741740        2      -5.65148881      -0.00000648       0.00001555      -0.02454311
-Pd      13.61028922      15.71580555      21.76741173        2      -5.65149137       0.00000458       0.00001316      -0.02450489
-Pd      16.33234709      15.71580568      21.76741417        2      -5.65148904      -0.00001124       0.00002061      -0.02452536
-Pd      19.05440283      15.71580885      21.76741250        2      -5.65148911       0.00001915      -0.00000516      -0.02451389
-Pd      21.77646485      15.71580777      21.76741607        2      -5.65149164      -0.00002163       0.00000179      -0.02453385
-Pd      24.49851993      15.71580716      21.76741452        2      -5.65148951       0.00000328       0.00000233      -0.02452697
-Pd       1.36102776      18.07317857      21.76741512        2      -5.65149099       0.00000615       0.00002046      -0.02454022
-Pd       4.08308429      18.07317667      21.76741460        2      -5.65149307       0.00002221       0.00001338      -0.02452554
-Pd       6.80514452      18.07318015      21.76741496        2      -5.65149421      -0.00001569      -0.00001939      -0.02453282
-Pd       9.52720242      18.07318021      21.76741425        2      -5.65148923      -0.00000417      -0.00000962      -0.02452245
-Pd      12.24925879      18.07317878      21.76741549        2      -5.65149250       0.00001261       0.00001255      -0.02453440
-Pd      14.97131824      18.07317932      21.76741286        2      -5.65149256      -0.00000876       0.00000915      -0.02451564
-Pd      17.69337507      18.07317940      21.76741443        2      -5.65148887      -0.00000300       0.00000457      -0.02452462
-Pd      20.41543214      18.07317775      21.76741295        2      -5.65149231       0.00001274       0.00001426      -0.02451121
-Pd      23.13749147      18.07318210      21.76741360        2      -5.65148872      -0.00001203      -0.00002265      -0.02451290
-Pd      25.85954836      18.07317820      21.76741485        2      -5.65148968       0.00000170       0.00000351      -0.02452954
-Pd      -0.00000025      20.43055165      21.76741414        2      -5.65149097       0.00001347      -0.00001013      -0.02453047
-Pd       2.72205831      20.43055313      21.76741344        2      -5.65149192      -0.00001144      -0.00003327      -0.02452364
-Pd       5.44411537      20.43055043      21.76741429        2      -5.65149137      -0.00000774      -0.00000812      -0.02452773
-Pd       8.16617309      20.43055075      21.76741623        2      -5.65148944       0.00000319      -0.00000552      -0.02455285
-Pd      10.88822999      20.43055409      21.76741281        2      -5.65149018       0.00000837      -0.00002879      -0.02453055
-Pd      13.61028741      20.43055102      21.76741273        2      -5.65149311       0.00001080      -0.00001540      -0.02451941
-Pd      16.33234880      20.43055401      21.76741616        2      -5.65148764      -0.00003269      -0.00004502      -0.02454935
-Pd      19.05440342      20.43055116      21.76741572        2      -5.65149010       0.00001948      -0.00001477      -0.02454210
-Pd      21.77646254      20.43055112      21.76741632        2      -5.65148922      -0.00001118      -0.00000696      -0.02453755
-Pd      24.49851931      20.43054999      21.76741555        2      -5.65148856       0.00000402       0.00000400      -0.02453007
-Pd       1.36102792      22.78792232      21.76741703        2      -5.65149023       0.00000686      -0.00001363      -0.02454862
-Pd       4.08308555      22.78792006      21.76741464        2      -5.65149227       0.00001141       0.00002061      -0.02453922
-Pd       6.80514596      22.78792178      21.76741754        2      -5.65148919      -0.00002170      -0.00000173      -0.02456189
-Pd       9.52720284      22.78792071      21.76741937        2      -5.65148781      -0.00000136       0.00001005      -0.02457742
-Pd      12.24926017      22.78792285      21.76741364        2      -5.65148954       0.00000230      -0.00001271      -0.02452790
-Pd      14.97131788      22.78792077      21.76741500        2      -5.65148914      -0.00001045       0.00001145      -0.02452127
-Pd      17.69337434      22.78791890      21.76741669        2      -5.65149087       0.00002175       0.00003626      -0.02455096
-Pd      20.41543405      22.78792101      21.76741303        2      -5.65149102      -0.00001214       0.00000386      -0.02452386
-Pd      23.13749010      22.78792304      21.76741395        2      -5.65148970       0.00000381      -0.00001285      -0.02452494
-Pd      25.85954775      22.78792080      21.76741185        2      -5.65149203       0.00000203       0.00000955      -0.02450632
-Pd       0.00000048       0.00000196      24.06571850        1      -5.26092108      -0.00001869      -0.00002978       0.01195820
-Pd       2.72205728       0.00000023      24.06571928        1      -5.26092081       0.00001102       0.00000213       0.01195088
-Pd       5.44411635      -0.00000091      24.06571797        1      -5.26092274      -0.00000700       0.00000697       0.01196095
-Pd       8.16617555       0.00000012      24.06571511        1      -5.26092210      -0.00001452       0.00000554       0.01198068
-Pd      10.88823230       0.00000167      24.06571648        1      -5.26092088       0.00000008      -0.00001132       0.01196355
-Pd      13.61028585      -0.00000241      24.06571695        1      -5.26092273       0.00002342       0.00001129       0.01196081
-Pd      16.33234576      -0.00000008      24.06571470        1      -5.26092337      -0.00000182       0.00000912       0.01196766
-Pd      19.05440528      -0.00000085      24.06571711        1      -5.26092289      -0.00000845      -0.00000127       0.01195811
-Pd      21.77646197       0.00000103      24.06571470        1      -5.26092186      -0.00000865      -0.00002151       0.01197122
-Pd      24.49851826       0.00000066      24.06571639        1      -5.26092224       0.00002508      -0.00001398       0.01195739
-Pd       1.36102910       2.35737041      24.06571656        1      -5.26092283       0.00000489       0.00001254       0.01196262
-Pd       4.08308773       2.35737176      24.06571728        1      -5.26092092      -0.00000896      -0.00000657       0.01195685
-Pd       6.80514627       2.35737071      24.06571795        1      -5.26092149      -0.00000475       0.00000649       0.01195350
-Pd       9.52720339       2.35736865      24.06571636        1      -5.26092261      -0.00000550       0.00001531       0.01195699
-Pd      12.24925755       2.35737123      24.06571606        1      -5.26092281       0.00001791       0.00000898       0.01195182
-Pd      14.97131743       2.35737231      24.06571692        1      -5.26092244      -0.00000837      -0.00001866       0.01196457
-Pd      17.69337421       2.35737155      24.06571718        1      -5.26092244       0.00001082      -0.00001163       0.01196508
-Pd      20.41543452       2.35737050      24.06571761        1      -5.26092241      -0.00001420       0.00001734       0.01196045
-Pd      23.13749177       2.35736701      24.06571868        1      -5.26092197      -0.00000160       0.00004577       0.01195188
-Pd      25.85954881       2.35736900      24.06571814        1      -5.26092153      -0.00000098       0.00002597       0.01194579
-Pd       0.00000095       4.71474233      24.06572015        1      -5.26092421      -0.00001267       0.00000332       0.01193245
-Pd       2.72205710       4.71474108      24.06571862        1      -5.26092189       0.00000161       0.00001534       0.01194032
-Pd       5.44411437       4.71474223      24.06571660        1      -5.26092035       0.00001418      -0.00000061       0.01195021
-Pd       8.16617582       4.71474098      24.06571926        1      -5.26092034      -0.00002889       0.00000728       0.01194455
-Pd      10.88822795       4.71474311      24.06571854        1      -5.26092142       0.00003160      -0.00000967       0.01194428
-Pd      13.61029028       4.71474102      24.06571687        1      -5.26092137      -0.00001840       0.00000017       0.01195808
-Pd      16.33234928       4.71474310      24.06571700        1      -5.26092099      -0.00001886       0.00000099       0.01195889
-Pd      19.05440354       4.71474289      24.06571687        1      -5.26092230       0.00001766      -0.00000946       0.01196017
-Pd      21.77646236       4.71474349      24.06571803        1      -5.26092145       0.00000598      -0.00002251       0.01194414
-Pd      24.49852021       4.71474171      24.06571713        1      -5.26092266       0.00000125      -0.00000313       0.01194158
-Pd       1.36102941       7.07211573      24.06571796        1      -5.26092227      -0.00000486      -0.00001794       0.01194484
-Pd       4.08308565       7.07211417      24.06571658        1      -5.26092166      -0.00000253      -0.00000749       0.01195768
-Pd       6.80514213       7.07211255      24.06571716        1      -5.26092165       0.00001701      -0.00000260       0.01195196
-Pd       9.52720386       7.07211226      24.06571913        1      -5.26092125      -0.00001519       0.00002305       0.01193913
-Pd      12.24926136       7.07211238      24.06571731        1      -5.26092217      -0.00001348       0.00001007       0.01195779
-Pd      14.97131655       7.07211143      24.06571702        1      -5.26092259       0.00001337       0.00001761       0.01195433
-Pd      17.69337515       7.07211305      24.06571808        1      -5.26092109      -0.00000432       0.00001717       0.01195012
-Pd      20.41543236       7.07211277      24.06571497        1      -5.26092075       0.00000227       0.00000199       0.01197497
-Pd      23.13749065       7.07211460      24.06572016        1      -5.26092063       0.00000674      -0.00000945       0.01194283
-Pd      25.85954747       7.07211525      24.06571742        1      -5.26092110       0.00001254      -0.00002075       0.01194413
-Pd      -0.00000015       9.42948504      24.06571921        1      -5.26092062      -0.00000010      -0.00000563       0.01195085
-Pd       2.72205751       9.42948808      24.06571627        1      -5.26092270       0.00000026      -0.00001836       0.01196089
-Pd       5.44411410       9.42948647      24.06571644        1      -5.26092313       0.00000714      -0.00000246       0.01196552
-Pd       8.16617190       9.42948450      24.06571566        1      -5.26092294       0.00001921       0.00000134       0.01195722
-Pd      10.88823286       9.42948784      24.06571773        1      -5.26092036      -0.00002944      -0.00003552       0.01195153
-Pd      13.61028619       9.42948592      24.06571809        1      -5.26092206       0.00002486      -0.00001877       0.01195333
-Pd      16.33234566       9.42948410      24.06571666        1      -5.26092184       0.00000711      -0.00000294       0.01195737
-Pd      19.05440752       9.42948557      24.06571747        1      -5.26092155      -0.00003841      -0.00000874       0.01195463
-Pd      21.77646057       9.42948423      24.06571770        1      -5.26092211       0.00002094       0.00000399       0.01196415
-Pd      24.49852044       9.42948363      24.06571838        1      -5.26092071      -0.00001022       0.00001016       0.01195190
-Pd       1.36102886      11.78685586      24.06571830        1      -5.26092267      -0.00000416      -0.00000022       0.01196339
-Pd       4.08308538      11.78685643      24.06571815        1      -5.26092204       0.00001080      -0.00000201       0.01195048
-Pd       6.80514442      11.78685724      24.06571804        1      -5.26092166      -0.00000569      -0.00002585       0.01196208
-Pd       9.52720232      11.78685986      24.06571580        1      -5.26092170      -0.00000884      -0.00003656       0.01197100
-Pd      12.24926003      11.78685491      24.06571583        1      -5.26092105      -0.00000010       0.00001529       0.01198088
-Pd      14.97131695      11.78685787      24.06571540        1      -5.26092165       0.00000924      -0.00001948       0.01197530
-Pd      17.69337487      11.78685486      24.06571698        1      -5.26092160       0.00000309      -0.00000541       0.01195765
-Pd      20.41543275      11.78685553      24.06571724        1      -5.26092219       0.00000561       0.00000338       0.01196518
-Pd      23.13749221      11.78685521      24.06571730        1      -5.26092146      -0.00000679      -0.00000161       0.01195761
-Pd      25.85954876      11.78685495      24.06571654        1      -5.26092193      -0.00000340       0.00000840       0.01196682
-Pd       0.00000033      14.14422767      24.06571447        1      -5.26092242      -0.00000332      -0.00001277       0.01198608
-Pd       2.72205719      14.14422475      24.06571531        1      -5.26092200       0.00000907       0.00001862       0.01196971
-Pd       5.44411698      14.14422529      24.06571469        1      -5.26092281      -0.00002118       0.00001306       0.01197874
-Pd       8.16617079      14.14422455      24.06571452        1      -5.26092205       0.00002354       0.00003261       0.01197628
-Pd      10.88822994      14.14422569      24.06571481        1      -5.26092204       0.00000469       0.00001994       0.01197461
-Pd      13.61028877      14.14422494      24.06571574        1      -5.26092196      -0.00000188       0.00002549       0.01196439
-Pd      16.33234777      14.14422766      24.06571577        1      -5.26092112      -0.00001232       0.00000317       0.01197655
-Pd      19.05440369      14.14422648      24.06571667        1      -5.26092210       0.00000521       0.00000771       0.01197167
-Pd      21.77646217      14.14422711      24.06571602        1      -5.26092273       0.00000534       0.00000306       0.01196942
-Pd      24.49852020      14.14422761      24.06571641        1      -5.26092131      -0.00000435      -0.00000762       0.01196429
-Pd       1.36102919      16.50159752      24.06571520        1      -5.26092230       0.00000967       0.00000403       0.01197469
-Pd       4.08308805      16.50159686      24.06571848        1      -5.26092112      -0.00000354       0.00000054       0.01194867
-Pd       6.80514555      16.50159727      24.06571728        1      -5.26092184      -0.00000543       0.00000927       0.01195037
-Pd       9.52720133      16.50159882      24.06571748        1      -5.26092065       0.00000178      -0.00001097       0.01195027
-Pd      12.24926142      16.50159802      24.06571366        1      -5.26092363      -0.00000508      -0.00000494       0.01197116
-Pd      14.97131743      16.50159967      24.06571719        1      -5.26092122      -0.00000002      -0.00001455       0.01194759
-Pd      17.69337515      16.50159793      24.06571665        1      -5.26092190       0.00000134       0.00000472       0.01195820
-Pd      20.41543196      16.50159760      24.06571765        1      -5.26092153       0.00001259       0.00000708       0.01194974
-Pd      23.13749346      16.50160048      24.06571544        1      -5.26092325      -0.00002459      -0.00000904       0.01196482
-Pd      25.85954911      16.50159711      24.06571714        1      -5.26092272      -0.00000084       0.00000663       0.01196656
-Pd      -0.00000182      18.85896905      24.06571497        1      -5.26092338       0.00001818       0.00000196       0.01196627
-Pd       2.72205713      18.85896695      24.06571786        1      -5.26092126       0.00000378       0.00001927       0.01194887
-Pd       5.44411626      18.85897062      24.06571890        1      -5.26092193      -0.00000978      -0.00001236       0.01194580
-Pd       8.16617323      18.85896985      24.06571780        1      -5.26092215       0.00000250      -0.00000371       0.01195905
-Pd      10.88823107      18.85896896      24.06571534        1      -5.26092301      -0.00000018       0.00000429       0.01196182
-Pd      13.61029010      18.85897029      24.06572003        1      -5.26092046      -0.00000759      -0.00000305       0.01193966
-Pd      16.33234680      18.85896873      24.06571592        1      -5.26092212      -0.00000076      -0.00000404       0.01196016
-Pd      19.05440403      18.85897028      24.06571683        1      -5.26092225       0.00000020      -0.00001499       0.01195782
-Pd      21.77646123      18.85897106      24.06571883        1      -5.26092050       0.00000925      -0.00001806       0.01194819
-Pd      24.49852072      18.85896765      24.06571872        1      -5.26092182      -0.00001650       0.00001231       0.01195670
-Pd       1.36102779      21.21634059      24.06571948        1      -5.26092144       0.00000090      -0.00000871       0.01194844
-Pd       4.08308620      21.21634117      24.06571737        1      -5.26092207      -0.00000349      -0.00001095       0.01195899
-Pd       6.80514480      21.21634077      24.06571457        1      -5.26092260      -0.00000521      -0.00000280       0.01197675
-Pd       9.52720232      21.21634242      24.06571508        1      -5.26092270       0.00000345      -0.00001986       0.01197967
-Pd      12.24925847      21.21633863      24.06571639        1      -5.26092199       0.00001385       0.00000681       0.01195688
-Pd      14.97131988      21.21633900      24.06571611        1      -5.26092168      -0.00003036       0.00000774       0.01196230
-Pd      17.69337348      21.21633933      24.06571567        1      -5.26092084       0.00002282       0.00000799       0.01198164
-Pd      20.41543403      21.21633846      24.06571508        1      -5.26092299      -0.00000724       0.00002320       0.01196985
-Pd      23.13749210      21.21634258      24.06571778        1      -5.26092193      -0.00000887      -0.00001465       0.01195877
-Pd      25.85954604      21.21633812      24.06571635        1      -5.26092163       0.00002112       0.00002942       0.01196139
-""")
-    atoms = read(f, format="extxyz")
-    assert isinstance(atoms, Atoms)
-    return atoms
